- Removed update loop from the renderer - Added update loop to Simulation - Some documentation - Removed glm::vec3 constant construction and replaced with constant multiplication.

+++ physics_engine/physics_engine/Object.cpp
@@ -1,6 +1,6 @@
 #include "Object.h"
 
-Object::Object(ObjectType type, double mass)
+Object::Object(ObjectType type, float mass)
     : mType(type)
 {
 	mPhysicsObj.setMass(mass);
